refactor(csv): consolidate CSV output functions and introduce ISO date parsing for week labels
