+++ utils/output/csv.js
@@ -1,3 +1,5 @@
+const { parseWeekLabelToStartEnd, toISODate } = require('../date');
+
 function formatWeeklyCsv(stats) {
   let csvOutput = 'Week Start,Week End,Name,Messages Sent,Messages Read,Average Read Time (minutes),Total Words, Sentiment, Sentiment_natural, Tone\n';
   for (const [week, weekStats] of Object.entries(stats)) {
@@ -15,54 +17,6 @@ function formatWeeklyCsv(stats) {
   return csvOutput;
 }
 
-// Accepts either:
-// 1) ISO range: YYYY-MM-DD – YYYY-MM-DD (em dash or hyphen)
-// 2) Short month range: Jan 01 - Jan 07, 2025 (as produced by getWeekString)
-function parseWeekLabelToStartEnd(label) {
-  if (!label || typeof label !== 'string') return { startISO: '', endISO: '' };
-  // Try ISO first (em dash or hyphen)
-  const isoMatch = label.match(/(\d{4}-\d{2}-\d{2})\s*[–-]\s*(\d{4}-\d{2}-\d{2})/);
-  if (isoMatch) {
-    return { startISO: isoMatch[1], endISO: isoMatch[2] };
-  }
-
-  // Try "Mon dd - Mon dd, yyyy"
-  const m = label.match(/^([A-Za-z]{3})\s+(\d{1,2})\s*-\s*([A-Za-z]{3})\s+(\d{1,2}),\s*(\d{4})$/);
-  if (!m) return { startISO: '', endISO: '' };
-
-  const monthToIndex = {
-    jan: 0, feb: 1, mar: 2, apr: 3, may: 4, jun: 5,
-    jul: 6, aug: 7, sep: 8, oct: 9, nov: 10, dec: 11,
-  };
-  const startMon = monthToIndex[m[1].toLowerCase()];
-  const startDay = Number(m[2]);
-  const endMon = monthToIndex[m[3].toLowerCase()];
-  const endDay = Number(m[4]);
-  const endYear = Number(m[5]);
-
-  // Determine start year (week can cross the year boundary)
-  let startYear = endYear;
-  if (startMon === 11 && endMon === 0) {
-    // Dec → Jan transition
-    startYear = endYear - 1;
-  }
-
-  const startDate = new Date(startYear, startMon, startDay);
-  const endDate = new Date(endYear, endMon, endDay);
-
-  return {
-    startISO: toISODate(startDate),
-    endISO: toISODate(endDate),
-  };
-}
-
-function toISODate(d) {
-  const yyyy = d.getFullYear();
-  const mm = String(d.getMonth() + 1).padStart(2, '0');
-  const dd = String(d.getDate()).padStart(2, '0');
-  return `${yyyy}-${mm}-${dd}`;
-}
-
 module.exports = { formatWeeklyCsv };
 
 // Build a second CSV focused on the global top 2 senders across all weeks
@@ -73,6 +27,7 @@ function formatWeeklyTop2Csv(stats) {
   const initialA = nameA.charAt(0).toUpperCase();
   const initialB = nameB.charAt(0).toUpperCase();
 
+  // Header A: Week Start first
   let csvOutput = `Week Start,Sent ${nameA},Sent ${nameB},Total Words ${nameA},Total Words ${nameB},Read Time ${nameA},Read Time ${nameB},Sentiment ${nameA},Sentiment ${nameB},Sentiment Natural ${nameA},Sentiment Natural ${nameB},Tone ${nameA},Tone ${nameB}` + "\n";
 
   const weeks = Object.keys(stats);
@@ -80,11 +35,11 @@ function formatWeeklyTop2Csv(stats) {
     const w = stats[week] || {};
     const a = w[nameA] || {};
     const b = w[nameB] || {};
-    const { startISO, endISO } = parseWeekLabelToStartEnd(week);
+    const { startISO } = parseWeekLabelToStartEnd(week);
     const toneA = (w[nameA] && Number.isFinite(Number(w[nameA].tone))) ? w[nameA].tone : 0;
     const toneB = (w[nameB] && Number.isFinite(Number(w[nameB].tone))) ? w[nameB].tone : 0;
     const row = [
-      startISO || '',
+      week || '',
       safeInt(a.messagesSent),
       safeInt(b.messagesSent),
       safeInt(a.totalWords),
